feat: 버튼 스위치, 유효성 검사 함수로 분리 feature/TODAK-213/update-testcase
