+++ domain/auth/components/signupSteps/SignupStepper.tsx
@@ -7,7 +7,7 @@ import { useSelector } from "react-redux";
 import { SIGNUP_STEP } from "../../constants";
 
 const SignupStepper: React.FC = () => {
-  const signupStep = useSelector((state: RootState) => state.signup.steps);
+  const signupStep = useSelector((state: RootState) => state.signup.step);
   return (
     <ol className="flex items-center w-full p-3 space-x-2 text-sm font-medium text-center text-gray-500 bg-white border border-gray-200 rounded-lg shadow-sm dark:text-gray-400 dark:bg-gray-800 dark:border-gray-700">
       <li
