+++ domain/auth/function/index.ts
@@ -1,0 +1,2 @@
+import handleSwitchSignupStep from "./validate";
+export { handleSwitchSignupStep };
